+++ Changelog.md
@@ -1,6 +1,21 @@
 CHANGELOG
 =========
 
+3.5.0
+-----
+
+2016/12/04
+
+### New Features
+
+Pass username in option.
+
+### Changes
+
+Detect failed authentication.
+Detect authentication expiration.
+
+
 3.4.0
 -----